+++ rust/otap-dataflow/crates/engine/src/extension/wrapper.rs
@@ -12,6 +12,7 @@
 //! [`local::extension`](crate::local::extension) and
 //! [`shared::extension`](crate::shared::extension).
 
+use crate::capability::factory::{LocalInstanceFactory, SharedInstanceFactory};
 use crate::channel_metrics::ChannelMetricsRegistry;
 use crate::channel_mode::{LocalMode, SharedMode, wrap_control_channel_metrics};
 use crate::config::ExtensionConfig;
@@ -30,7 +31,6 @@ use otap_df_config::ExtensionId;
 use otap_df_config::extension::ExtensionUserConfig;
 use otap_df_telemetry::otel_debug;
 use otap_df_telemetry::reporter::MetricsReporter;
-use std::any::Any;
 use std::future::Future;
 use std::pin::Pin;
 use std::sync::Arc;
@@ -224,128 +224,6 @@ pub enum ExtensionLifecycle<E, R> {
     },
     /// Passive extension — capabilities only, no task spawned.
     Passive,
-}
-
-// ── InstanceFactory ──────────────────────────────────────────────────────────
-
-/// Produces instances of a shared extension's concrete type for capability
-/// consumers.
-///
-/// The wrapper is type-erased over the extension's concrete type `E`. When
-/// the engine wires a consumer's capability binding, the generated
-/// registration glue (see `extension_capabilities!`) downcasts the returned
-/// `Box<dyn Any + Send>` back to `E` and wraps it in the requested
-/// capability trait object.
-///
-/// The instance policy chosen at the builder (`.cloned(...)` vs
-/// `.fresh(...)`) is baked into the stored closure:
-/// - **Cloned** — the closure captures a prototype `E: Clone`
-///   and returns `e.clone()` on each call.
-/// - **Fresh** — the closure captures the user-supplied
-///   `Fn() -> E` and invokes it on each call.
-///
-/// `SharedInstanceFactory` is [`Clone`]: one extension may provide
-/// multiple capabilities, and each capability registration needs its
-/// own copy of the factory for its produce closure.
-pub struct SharedInstanceFactory {
-    produce: Box<dyn SharedFnClone>,
-}
-
-/// Object-safe `Fn + Clone` helper for [`SharedInstanceFactory`].
-///
-/// `Box<dyn Fn>` is not `Clone`, so we thread cloning through an
-/// object-safe `clone_box` method. The blanket impl below covers any
-/// closure that is `Fn + Send + Clone + 'static`.
-#[doc(hidden)]
-pub trait SharedFnClone: Fn() -> Box<dyn Any + Send> + Send {
-    fn clone_box(&self) -> Box<dyn SharedFnClone>;
-}
-
-impl<F> SharedFnClone for F
-where
-    F: Fn() -> Box<dyn Any + Send> + Send + Clone + 'static,
-{
-    fn clone_box(&self) -> Box<dyn SharedFnClone> {
-        Box::new(self.clone())
-    }
-}
-
-impl SharedInstanceFactory {
-    /// Construct a factory from a closure producing type-erased instances.
-    #[must_use]
-    pub fn new<F>(produce: F) -> Self
-    where
-        F: Fn() -> Box<dyn Any + Send> + Send + Clone + 'static,
-    {
-        SharedInstanceFactory {
-            produce: Box::new(produce),
-        }
-    }
-
-    /// Produce a fresh instance as `Box<dyn Any + Send>`.
-    #[must_use]
-    pub fn produce(&self) -> Box<dyn Any + Send> {
-        (self.produce)()
-    }
-}
-
-impl Clone for SharedInstanceFactory {
-    fn clone(&self) -> Self {
-        SharedInstanceFactory {
-            produce: self.produce.clone_box(),
-        }
-    }
-}
-
-/// Produces instances of a local (!Send) extension's concrete type for
-/// capability consumers. See [`SharedInstanceFactory`] for background.
-///
-/// `LocalInstanceFactory` is [`Clone`] for the same reason as its
-/// shared counterpart.
-pub struct LocalInstanceFactory {
-    produce: Box<dyn LocalFnClone>,
-}
-
-/// Object-safe `Fn + Clone` helper for [`LocalInstanceFactory`].
-#[doc(hidden)]
-pub trait LocalFnClone: Fn() -> std::rc::Rc<dyn Any> {
-    fn clone_box(&self) -> Box<dyn LocalFnClone>;
-}
-
-impl<F> LocalFnClone for F
-where
-    F: Fn() -> std::rc::Rc<dyn Any> + Clone + 'static,
-{
-    fn clone_box(&self) -> Box<dyn LocalFnClone> {
-        Box::new(self.clone())
-    }
-}
-
-impl LocalInstanceFactory {
-    /// Construct a factory from a closure producing type-erased instances.
-    #[must_use]
-    pub fn new<F>(produce: F) -> Self
-    where
-        F: Fn() -> std::rc::Rc<dyn Any> + Clone + 'static,
-    {
-        LocalInstanceFactory {
-            produce: Box::new(produce),
-        }
-    }
-
-    /// Produce an instance as `Rc<dyn Any>`.
-    #[must_use]
-    pub fn produce(&self) -> std::rc::Rc<dyn Any> {
-        (self.produce)()
-    }
-}
-
-impl Clone for LocalInstanceFactory {
-    fn clone(&self) -> Self {
-        LocalInstanceFactory {
-            produce: self.produce.clone_box(),
-        }
-    }
 }
 
 // ── ExtensionWrapper ────────────────────────────────────────────────────────
